apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dream-central-api"
version = "1.0.0"
description = "FastAPI backend service for Flow Central Storage"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Flow Central Team" }]
dependencies = [
  "fastapi>=0.111,<0.112",
  "uvicorn[standard]>=0.30,<0.31",
  "pydantic-settings>=2.4,<2.5",
  "sqlalchemy>=2.0,<2.1",
  "alembic>=1.13,<1.14",
  "psycopg[binary]>=3.2,<3.3",
  "minio>=7.2,<7.3",
  "bcrypt>=4.0,<5.0",
  "httpx>=0.27,<0.28",
  "edge-tts>=7.2,<8.0",
  "arq>=0.26,<0.27",
  "redis>=5.0,<6.0",
  "pymupdf>=1.24,<2.0",
  "python-docx>=1.1,<2.0",
  "cefrpy>=1.0,<2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<8.3",
  "httpx>=0.27,<0.28",
  "pytest-asyncio>=0.23,<0.24"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
where = ["app"]
include = ["app*"]

[tool.setuptools.package-dir]
"" = "app"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
per-file-ignores = {"alembic/env.py" = ["E402"]}

[tool.ruff.format]
quote-style = "double"
